requirements.txt
git+https://github.com/openai/swarm.git
python-dotenv==0.19.0
openai>=1.33.0
streamlit>=1.32.0